+++ CHANGELOG.md
@@ -55,6 +55,11 @@ Support without rootElement
 * Register fastboot transform for apps to wrap non compatible asserts (PR#470) (@kratiahuja)
 * Move fastboot-transform and fs-extra to depedencies in package.json (@SergeAstapov)
 
+## Unreleased
+
+### Added
+* allow reading sandboxGlobals from app's config/environment
+
 ## 1.1.0
 
 * Bumping `fastboot-express-middleware` to 1.1.0